more updates

+++ routes/categories.routes.js
@@ -1,4 +1,5 @@
 const { Router } = require('express');
+const { Types: { ObjectId: { isValid } } } = require('mongoose');
 const { Category, validateCategory } = require('../models/category.model');
 const { auth } = require('../middlewares/auth.middleware');
 const { isAdmin } = require('../middlewares/isAdmin.middleware');
@@ -10,6 +11,7 @@ router.get('/', async (req, res) => {
 })
 
 router.get('/:id', async (req, res) => {
+  if (!isValid(req.params.id)) return res.status(404).send('Yaroqsiz id');
   const category = await Category.findById(req.params.id);
   if (!category) return res.status(404).send('Kategoriya topilmadi');
   res.send(category);
@@ -27,6 +29,7 @@ router.post('/', auth, async (req, res) => {
 })
 
 router.put('/:id', auth, async (req, res) => {
+  if (!isValid(req.params.id)) return res.status(404).send('Yaroqsiz id');
   const { error } = validateCategory(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
@@ -38,6 +41,7 @@ router.put('/:id', auth, async (req, res) => {
 })
 
 router.delete('/:id', [auth, isAdmin], async (req, res) => {
+  if (!isValid(req.params.id)) return res.status(404).send('Yaroqsiz id');
   let category = await Category.findByIdAndRemove(req.params.id);
   if (!category) return res.status(404).send('Kategoriya topilmadi');
   res.send(category);
